+++ lib/mongoid/compatibility.rb
@@ -1,0 +1,4 @@
+require 'mongoid'
+require 'mongoid/compatibility/self'
+require 'mongoid/compatibility/version'
+require 'mongoid/compatibility/object_id'