+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.1.1 (Next)
 
+* [#1](https://github.com/dblock/mongoid-compatibility/issues/1): Require `mongoid/compatibility` - [@dblock](https://github.com/dblock).
 * Your contribution here.
 
 ### 0.1.0 (2015/09/17)